feat(solutions-form): adds solution form
